+++ vtr_storage/include/vtr_storage/data_stream_writer.inl
@@ -1,22 +1,24 @@
-// description
 #pragma once
 
-#include "vtr_storage/DataStreamWriter.hpp"
+#include "vtr_storage/data_stream_writer.hpp"
 
 namespace vtr {
 namespace storage {
 
-template<typename MessageType>
-DataStreamWriter<MessageType>::DataStreamWriter(const std::string &data_directory_string,
-                                   const std::string &stream_name, bool append)
+template <typename MessageType>
+DataStreamWriter<MessageType>::DataStreamWriter(
+    const std::string &data_directory_string, const std::string &stream_name,
+    bool append)
     : DataStreamWriterBase(data_directory_string, stream_name, append) {
   tm_ = createTopicMetadata();
 }
 
-template<typename MessageType>
-DataStreamWriter<MessageType>::~DataStreamWriter() { close(); }
+template <typename MessageType>
+DataStreamWriter<MessageType>::~DataStreamWriter() {
+  close();
+}
 
-template<typename MessageType>
+template <typename MessageType>
 void DataStreamWriter<MessageType>::open() {
   if (!this->opened_) {
     writer_ = std::make_shared<SequentialAppendWriter>();
@@ -26,14 +28,15 @@ void DataStreamWriter<MessageType>::open() {
   }
 }
 
-template<typename MessageType>
+template <typename MessageType>
 void DataStreamWriter<MessageType>::close() {
   writer_.reset();
   this->opened_ = false;
 }
 
-template<typename MessageType>
-rosbag2_storage::TopicMetadata DataStreamWriter<MessageType>::createTopicMetadata() {
+template <typename MessageType>
+rosbag2_storage::TopicMetadata
+DataStreamWriter<MessageType>::createTopicMetadata() {
   // ToDo: create topic based on topic name
   rosbag2_storage::TopicMetadata tm;
   tm.name = "/my/test/topic";
@@ -42,12 +45,13 @@ rosbag2_storage::TopicMetadata DataStreamWriter<MessageType>::createTopicMetadat
   return tm;
 }
 
-template<typename MessageType>
+template <typename MessageType>
 int32_t DataStreamWriter<MessageType>::write(const std::any &anytype_message) {
   if (!this->opened_) open();
   try {
     auto message = std::any_cast<MessageType>(anytype_message);
-    auto bag_message = std::make_shared<rosbag2_storage::SerializedBagMessage>();
+    auto bag_message =
+        std::make_shared<rosbag2_storage::SerializedBagMessage>();
     auto ret = rcutils_system_time_now(&bag_message->time_stamp);
     if (ret != RCL_RET_OK) {
       throw std::runtime_error("couldn't assign time rosbag message");
@@ -62,10 +66,11 @@ int32_t DataStreamWriter<MessageType>::write(const std::any &anytype_message) {
 
     writer_->write(bag_message);
     return writer_->get_last_inserted_id();
-  } catch (const std::bad_any_cast& e) {
-      std::stringstream ss;
-      ss << "Any cast failed in writing data in DataStreamWriter. Error: " << e.what();
-      throw std::runtime_error(ss.str());
+  } catch (const std::bad_any_cast &e) {
+    std::stringstream ss;
+    ss << "Any cast failed in writing data in DataStreamWriter. Error: "
+       << e.what();
+    throw std::runtime_error(ss.str());
   }
 }
