+++ vtr_storage/include/vtr_storage/random_access_reader.hpp
