+++ vtr_storage/include/vtr_storage/sequential_append_writer.hpp
